+++ web/src/components/LoginBox/index.tsx
@@ -1,46 +1,13 @@
-import { useEffect } from 'react'
+import { useContext } from 'react'
 import { VscGithubInverted } from 'react-icons/vsc'
-import { api } from '../../services/api'
+import { AuthContext } from '../../contexts/auth'
 import styles from './styles.module.scss'
 
-type AuthResponse = {
-    token: string;
-    user: {
-        id: string;
-        avatar_url: string;
-        name: string;
-        login: string;
-    }
-}
 
 export function LoginBox(){
-    const signInUrl = `https://github.com/login/oauth/authorize?scope=user&client_id=41be629ed0a2f9eb4975`
+    const {signInUrl, user} = useContext(AuthContext);
+    console.log(user)
 
-    async function signIn(githubCode: string){
-        const response = await api.post<AuthResponse>('authenticate', {
-            code: githubCode,
-        })
-
-        const { token, user } = response.data;
-
-        localStorage.setItem('@dowhile:token', token);
-
-        console.log(user)
-    }
-
-    useEffect(() => {
-        const url = window.location.href;
-        const hasGithubCode = url.includes('?code=');
-
-        if(hasGithubCode) {
-            const [urlWithoutCode, gitHubCode] = url.split('?code=');
-
-            window.history.pushState({}, '', urlWithoutCode);
-
-            signIn(gitHubCode);
-        }
-    }, [])
-''
     return(
         <div className={styles.loginBoxWrapper}>
             <strong>Entre e compartilhe sua mensagem</strong>
