+++ web/src/contexts/auth.tsx
@@ -1,14 +1,86 @@
-import { createContext, ReactNode } from "react";
+import { createContext, ReactNode, useEffect, useState } from "react";
+import { api } from "../services/api";
 
-const AuthContext = createContext(null);
+type User = {
+    id: string;
+    name: string;
+    login: string;
+    avatar_url: string;
+}
+
+type AuthContextData = {
+    user: User | null;
+    signInUrl: string;
+    signOut: () => void;
+}
+
+export const AuthContext = createContext({} as AuthContextData);
 
 type AuthProvider = {
     children: ReactNode;
 }
 
+type AuthResponse = {
+    token: string;
+    user: {
+        id: string;
+        avatar_url: string;
+        name: string;
+        login: string;
+    }
+}
+
+
 export function AuthProvider(props: AuthProvider){
+
+    const[user, setUser] = useState<User | null>(null)
+
+    const signInUrl = `https://github.com/login/oauth/authorize?scope=user&client_id=41be629ed0a2f9eb4975`
+
+    async function signIn(githubCode: string){
+        const response = await api.post<AuthResponse>('authenticate', {
+            code: githubCode,
+        })
+
+        const { token, user } = response.data;
+
+        localStorage.setItem('@dowhile:token', token);
+
+        setUser(user);
+    }
+
+    function signOut(){
+        setUser(null);
+        localStorage.removeItem('@dowhile:token');
+    }
+
+    useEffect(() => {
+        const token = localStorage.getItem('@dowhile:token');
+
+        if(token) {
+            api.defaults.headers.common.authorization = `Bearer ${token}`;
+
+            api.get<User>('profile').then(response => {
+                setUser(response.data)
+            })
+        }
+    }, [])
+
+    useEffect(() => {
+        const url = window.location.href;
+        const hasGithubCode = url.includes('?code=');
+
+        if(hasGithubCode) {
+            const [urlWithoutCode, gitHubCode] = url.split('?code=');
+
+            window.history.pushState({}, '', urlWithoutCode);
+
+            signIn(gitHubCode);
+        }
+    }, [])
+
     return (
-        <AuthContext.Provider value={null}>
+        <AuthContext.Provider value={{ signInUrl, user, signOut }}>
             {props.children}
         </AuthContext.Provider>
     )
